Move innerListRowsFrom ifNullString ifNullUUID UUIDFromSqlString to core
